fixed 6ba17fc from https://gitee.com/w253227059/kernel_liteos_a/pulls/853 fix: kora工具检视问题修复 【背景】kora工具检视问题修复
【修改方案】
1，对部分代码问题进行整改

【影响】
对现有的产品编译不会有影响。

+++ kernel/extended/hilog/los_hilog.c
@@ -148,7 +148,7 @@ static int HiLogBufferCopy(unsigned char *dst, unsigned dstLen, const unsigned c
 
 static int HiLogReadRingBuffer(unsigned char *buffer, size_t bufLen)
 {
-    size_t retval;
+    int retval;
     size_t bufLeft = HILOG_BUFFER - g_hiLogDev.headOffset;
     if (bufLeft > bufLen) {
         retval = HiLogBufferCopy(buffer, bufLen, HiLogBufferHead(), bufLen);
@@ -165,7 +165,7 @@ static int HiLogReadRingBuffer(unsigned char *buffer, size_t bufLen)
 
 static ssize_t HiLogRead(struct file *filep, char *buffer, size_t bufLen)
 {
-    size_t retval;
+    int retval;
     struct HiLogEntry header;
 
     (void)filep;
@@ -209,7 +209,7 @@ out:
         g_hiLogDev.count = 0;
     }
     (VOID)LOS_MuxRelease(&g_hiLogDev.mtx);
-    return retval;
+    return (ssize_t)retval;
 }
 
 static int HiLogWriteRingBuffer(unsigned char *buffer, size_t bufLen)
